snowflakes are 64bit

+++ discordprimatives/snowflakes.go
@@ -1,4 +1,4 @@
 package discordprimatives
 
 //Snowflake documented at https://discord.com/developers/docs/reference#snowflakes
-type Snowflake uint32 //todo: helper methods
+type Snowflake uint64 //todo: helper methods
